+++ src/modules/entry/entries-repository.ts
