+++ src/modules/journal/journals-dtos.ts
@@ -1,4 +1,4 @@
-import {OrderOption} from "../utils/order-option";
+import {OrderOption} from "../../utils/order-option";
 import {Journal, JournalSortOption} from "./journal";
 
 export type GetJournalDTO = {id: string};
